login/logout/register/middlewares

+++ routes/web.php
@@ -2,6 +2,7 @@
 
 use App\models\Listing;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
 use App\Http\Controllers\ListingController;
 
 /*
@@ -30,26 +31,39 @@ Route::get('/', [ListingController::class,'index']);
 
 
 // show create form
-Route::get('/listings/create',[ListingController::class,'create']);
+Route::get('/listings/create',[ListingController::class,'create'])->middleware('auth');
 
 
 // store listings data
-Route::post('/listings', [ListingController::class, 'store'] );
+Route::post('/listings', [ListingController::class, 'store'] )->middleware('auth');
 
 
 //show edit form
-Route::get('listings/{listing}/edit',[ListingController::class,'edit']);
+Route::get('listings/{listing}/edit',[ListingController::class,'edit'])->middleware('auth');
 
 // edit submit to update
-Route::put('/listings/{listing}',[ListingController::class,'update']);
+Route::put('/listings/{listing}',[ListingController::class,'update'])->middleware('auth');
 
 // edit submit to update
 // Route::put('/listings/{listing}',[ListingController::class,'update']);
 
 // delete
-Route::delete('/listings/{listing}',[ListingController::class,'destroy']);
+Route::delete('/listings/{listing}',[ListingController::class,'destroy'])->middleware('auth');
 
 // single listings
-Route::get('/listings/{listing}', [ListingController::class, 'show'] );
+Route::get('/listings/{listing}', [ListingController::class, 'show'] )->middleware('auth');
 
+// register
+Route::get('/register',[UserController::class,'create']);
 
+// create new user
+Route::post('/users',[UserController::class,'store'])->middleware('auth');
+
+// logout users
+Route::post('/logout',[UserController::class,'logout'])->middleware('auth');
+
+// login
+Route::get('/login',[UserController::class,'login'])->name('login');
+
+// 
+Route::post('/users/authenticate',[UserController::class,'authenticate']);
